requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg[binary]>=3.2.2
python-decouple==3.8
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
django-filter==23.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
requests==2.31.0
django-cors-headers==4.3.1

